Add support for setting a custom role when establishing a connection
