+++ src/Database/PostgreSQL/PQTypes/Internal/State.hs
@@ -1,9 +1,13 @@
 -- | Definition of internal DBT state.
-module Database.PostgreSQL.PQTypes.Internal.State (
-    DBState(..)
+module Database.PostgreSQL.PQTypes.Internal.State
+  ( DBState(..)
+  , updateStateWith
   ) where
 
+import Foreign.ForeignPtr
+
 import Database.PostgreSQL.PQTypes.FromRow
+import Database.PostgreSQL.PQTypes.Internal.C.Types
 import Database.PostgreSQL.PQTypes.Internal.Connection
 import Database.PostgreSQL.PQTypes.Internal.QueryResult
 import Database.PostgreSQL.PQTypes.SQL.Class
@@ -23,4 +27,14 @@ data DBState m = DBState
   , dbRecordLastQuery     :: !Bool
     -- | Current query result.
   , dbQueryResult         :: !(forall row. FromRow row => Maybe (QueryResult row))
+  }
+
+updateStateWith :: IsSQL sql => DBState m -> sql -> ForeignPtr PGresult -> DBState m
+updateStateWith st sql res = st
+  { dbLastQuery = if dbRecordLastQuery st then SomeSQL sql else dbLastQuery st
+  , dbQueryResult = Just QueryResult
+    { qrSQL = SomeSQL sql
+    , qrResult = res
+    , qrFromRow = id
+    }
   }
